requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
moviepy==1.0.3
openai-whisper==20240314
opencv-python==4.8.1.78
mediapipe==0.10.8
qtawesome==1.3.1
Pillow==10.0.0
numpy==1.24.3
scipy==1.11.3